Modified the get API details test alone with it's associated files
Changes included
- Correcting the test case APIVerificationTest.java
- Adding comments to the java classes
- Fixing the pom.xml file
- Adding log4j2.xml file
- Creating a new test suite file TestSuite.xml
