+++ src/main/java/com/assurity/common/Constants.java
@@ -1,0 +1,13 @@
+package com.assurity.common;
+
+/**
+ * <h1>Constants Class</h1>
+ * This class stores all the test data util methods
+ *
+ * @author Ramith Wanniarachchi
+ */
+public class Constants {
+    public static final String GET_ITEM_URL = "https://api.tmsandbox.co.nz/v1/Categories/<ITEM_NUMBER>/Details.json?catalogue=<CATALOGUE>";
+    public static final String ITEM_TEXT = "<ITEM_NUMBER>";
+    public static final String CATALOGUE_TEXT = "<CATALOGUE>";
+}
