--- src/main/java/com/assurityconsulting/common/Constants.java
@@ -1,7 +1,0 @@
-package com.assurityconsulting.common;
-
-public class Constants {
-    public static final String GET_ITEM_URL = "https://api.tmsandbox.co.nz/v1/Categories/<ITEM_NUMBER>/Details.json?catalogue=<CATALOGUE>";
-    public static final String ITEM_TEXT = "<ITEM_NUMBER>";
-    public static final String CATALOGUE_TEXT = "<CATALOGUE>";
-}
