+++ src/test/testSuites/TestSuite.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
+<suite name="AssurityTestSuite">
+    <test name="ItemGetAPITest">
+        <classes>
+            <class name="com.assurity.itemapi.APIVerificationTest"/>
+        </classes>
+    </test>
+</suite>
